refactor: Move exam creation to socket
- remove integration of Classroom model and `createExam` controller function in `examController.js`.
- integrate `Classroom` model and `createExam` function in `socket.exam.js`.
- refactor error handling to throw errors on invalid data in `createExam` in `socket.exam.js`.
- change props extracted from client-passed data from `userId` to `senderId` , and extract `quizId` from `data` instead of params in `createExam` in `socket.exam.js`.
- add `exam created` event emission with payload exam on successfully created exams in `createExam` in `socket.exam.js`.
